feat: accumulator v2 (wip 2)
