+++ runtime/Cargo.toml
@@ -35,6 +35,7 @@ num-traits = { version = "0.2" }
 num_cpus = "1.13.1"
 once_cell = "1.12.0"
 ouroboros = "0.15.0"
+pyth-oracle = { path = "../../pyth-client/program/rust" }
 pythnet-sdk = { git = "https://github.com/pyth-network/pyth-crosschain", version = "1.13.6", rev = "e670f57f89b05398ca352e4accb1e32724a8e1b4" }
 rand = "0.7.0"
 rayon = "1.5.3"
@@ -59,9 +60,11 @@ solana-zk-token-sdk = { path = "../zk-token-sdk", version = "=1.14.173" }
 strum = { version = "0.24", features = ["derive"] }
 strum_macros = "0.24"
 symlink = "0.1.0"
+syn = "=2.0.15"
 tar = "0.4.38"
 tempfile = "3.3.0"
-thiserror = "1.0"
+thiserror-impl = "=1.0.40"
+thiserror = "=1.0.40"
 zstd = "0.11.2"
 
 [lib]
