+++ sdk/program/src/account_info.rs
@@ -184,7 +184,6 @@ impl<'a> AccountInfo<'a> {
     pub fn assign(&self, new_owner: &Pubkey) {
         // Set the non-mut owner field
         unsafe {
-            #[allow(invalid_reference_casting)]
             std::ptr::write_volatile(
                 self.owner as *const Pubkey as *mut [u8; 32],
                 new_owner.to_bytes(),
